changing ints to float

+++ collated.go
@@ -17,8 +17,8 @@ type CollatedMetrics struct {
 	collateMax                     int
 	quit, running                  chan bool
 	collateCounters, collateGauges chan map[string]interface{}
-	counters, gauges               map[string]chan int64
-	customCounters, customGauges   map[string]chan map[string]int64
+	counters, gauges               map[string]chan float64
+	customCounters, customGauges   map[string]chan map[string]float64
 	httpClient                     *http.Client
 }
 
@@ -33,9 +33,9 @@ func NewCollatedMetrics(user, token, source string,
 		make(chan bool), make(chan bool),
 		make(chan map[string]interface{}, collateMax),
 		make(chan map[string]interface{}, collateMax),
-		make(map[string]chan int64), make(map[string]chan int64),
-		make(map[string]chan map[string]int64),
-		make(map[string]chan map[string]int64),
+		make(map[string]chan float64), make(map[string]chan float64),
+		make(map[string]chan map[string]float64),
+		make(map[string]chan map[string]float64),
 		&http.Client{Transport: http.DefaultTransport},
 	}
 
@@ -125,7 +125,7 @@ func (m *CollatedMetrics) Close() {
 }
 
 // Get (possibly by creating) a counter channel by the given name.
-func (m *CollatedMetrics) GetCounter(name string) chan int64 {
+func (m *CollatedMetrics) GetCounter(name string) chan float64 {
 	ch, ok := m.counters[name]
 	if ok {
 		return ch
@@ -134,7 +134,7 @@ func (m *CollatedMetrics) GetCounter(name string) chan int64 {
 }
 
 // Get (possibly by creating) a custom counter channel by the given name.
-func (m *CollatedMetrics) GetCustomCounter(name string) chan map[string]int64 {
+func (m *CollatedMetrics) GetCustomCounter(name string) chan map[string]float64 {
 	ch, ok := m.customCounters[name]
 	if ok {
 		return ch
@@ -143,7 +143,7 @@ func (m *CollatedMetrics) GetCustomCounter(name string) chan map[string]int64 {
 }
 
 // Get (possibly by creating) a custom gauge channel by the given name.
-func (m *CollatedMetrics) GetCustomGauge(name string) chan map[string]int64 {
+func (m *CollatedMetrics) GetCustomGauge(name string) chan map[string]float64 {
 	ch, ok := m.customGauges[name]
 	if ok {
 		return ch
@@ -152,7 +152,7 @@ func (m *CollatedMetrics) GetCustomGauge(name string) chan map[string]int64 {
 }
 
 // Get (possibly by creating) a gauge channel by the given name.
-func (m *CollatedMetrics) GetGauge(name string) chan int64 {
+func (m *CollatedMetrics) GetGauge(name string) chan float64 {
 	ch, ok := m.gauges[name]
 	if ok {
 		return ch
@@ -161,32 +161,32 @@ func (m *CollatedMetrics) GetGauge(name string) chan int64 {
 }
 
 // Create a counter channel by the given name.
-func (m *CollatedMetrics) NewCounter(name string) chan int64 {
-	ch := make(chan int64)
+func (m *CollatedMetrics) NewCounter(name string) chan float64 {
+	ch := make(chan float64)
 	m.counters[name] = ch
 	go m.newMetric(name, ch, m.collateCounters)
 	return ch
 }
 
 // Create a custom counter channel by the given name.
-func (m *CollatedMetrics) NewCustomCounter(name string) chan map[string]int64 {
-	ch := make(chan map[string]int64)
+func (m *CollatedMetrics) NewCustomCounter(name string) chan map[string]float64 {
+	ch := make(chan map[string]float64)
 	m.customCounters[name] = ch
 	go m.newMetric(name, ch, m.collateCounters)
 	return ch
 }
 
 // Create a custom gauge channel by the given name.
-func (m *CollatedMetrics) NewCustomGauge(name string) chan map[string]int64 {
-	ch := make(chan map[string]int64)
+func (m *CollatedMetrics) NewCustomGauge(name string) chan map[string]float64 {
+	ch := make(chan map[string]float64)
 	m.customGauges[name] = ch
 	go m.newMetric(name, ch, m.collateGauges)
 	return ch
 }
 
 // Create a gauge channel by the given name.
-func (m *CollatedMetrics) NewGauge(name string) chan int64 {
-	ch := make(chan int64)
+func (m *CollatedMetrics) NewGauge(name string) chan float64 {
+	ch := make(chan float64)
 	m.gauges[name] = ch
 	go m.newMetric(name, ch, m.collateGauges)
 	return ch
