+++ simple.go
@@ -23,8 +23,8 @@ var uaString = func() string {
 type SimpleMetrics struct {
 	user, token, source          string
 	quit, running                chan bool
-	counters, gauges             map[string]chan int64
-	customCounters, customGauges map[string]chan map[string]int64
+	counters, gauges             map[string]chan float64
+	customCounters, customGauges map[string]chan map[string]float64
 	httpClient                   *http.Client
 }
 
@@ -34,9 +34,9 @@ func NewSimpleMetrics(user, token, source string) Metrics {
 	m := &SimpleMetrics{
 		user, token, source,
 		make(chan bool), make(chan bool),
-		make(map[string]chan int64), make(map[string]chan int64),
-		make(map[string]chan map[string]int64),
-		make(map[string]chan map[string]int64),
+		make(map[string]chan float64), make(map[string]chan float64),
+		make(map[string]chan map[string]float64),
+		make(map[string]chan map[string]float64),
 		&http.Client{Transport: http.DefaultTransport},
 	}
 
@@ -88,7 +88,7 @@ func (m *SimpleMetrics) Close() {
 }
 
 // Get (possibly by creating) a counter channel by the given name.
-func (m *SimpleMetrics) GetCounter(name string) chan int64 {
+func (m *SimpleMetrics) GetCounter(name string) chan float64 {
 	ch, ok := m.counters[name]
 	if ok {
 		return ch
@@ -97,7 +97,7 @@ func (m *SimpleMetrics) GetCounter(name string) chan int64 {
 }
 
 // Get (possibly by creating) a custom counter channel by the given name.
-func (m *SimpleMetrics) GetCustomCounter(name string) chan map[string]int64 {
+func (m *SimpleMetrics) GetCustomCounter(name string) chan map[string]float64 {
 	ch, ok := m.customCounters[name]
 	if ok {
 		return ch
@@ -106,7 +106,7 @@ func (m *SimpleMetrics) GetCustomCounter(name string) chan map[string]int64 {
 }
 
 // Get (possibly by creating) a custom gauge channel by the given name.
-func (m *SimpleMetrics) GetCustomGauge(name string) chan map[string]int64 {
+func (m *SimpleMetrics) GetCustomGauge(name string) chan map[string]float64 {
 	ch, ok := m.customGauges[name]
 	if ok {
 		return ch
@@ -115,7 +115,7 @@ func (m *SimpleMetrics) GetCustomGauge(name string) chan map[string]int64 {
 }
 
 // Get (possibly by creating) a gauge channel by the given name.
-func (m *SimpleMetrics) GetGauge(name string) chan int64 {
+func (m *SimpleMetrics) GetGauge(name string) chan float64 {
 	ch, ok := m.gauges[name]
 	if ok {
 		return ch
@@ -124,32 +124,32 @@ func (m *SimpleMetrics) GetGauge(name string) chan int64 {
 }
 
 // Create a counter channel by the given name.
-func (m *SimpleMetrics) NewCounter(name string) chan int64 {
-	ch := make(chan int64)
+func (m *SimpleMetrics) NewCounter(name string) chan float64 {
+	ch := make(chan float64)
 	m.counters[name] = ch
 	go m.newMetric("counters", name, ch)
 	return ch
 }
 
 // Create a custom counter channel by the given name.
-func (m *SimpleMetrics) NewCustomCounter(name string) chan map[string]int64 {
-	ch := make(chan map[string]int64)
+func (m *SimpleMetrics) NewCustomCounter(name string) chan map[string]float64 {
+	ch := make(chan map[string]float64)
 	m.customCounters[name] = ch
 	go m.newMetric("counters", name, ch)
 	return ch
 }
 
 // Create a custom gauge channel by the given name.
-func (m *SimpleMetrics) NewCustomGauge(name string) chan map[string]int64 {
-	ch := make(chan map[string]int64)
+func (m *SimpleMetrics) NewCustomGauge(name string) chan map[string]float64 {
+	ch := make(chan map[string]float64)
 	m.customGauges[name] = ch
 	go m.newMetric("gauges", name, ch)
 	return ch
 }
 
 // Create a gauge channel by the given name.
-func (m *SimpleMetrics) NewGauge(name string) chan int64 {
-	ch := make(chan int64)
+func (m *SimpleMetrics) NewGauge(name string) chan float64 {
+	ch := make(chan float64)
 	m.gauges[name] = ch
 	go m.newMetric("gauges", name, ch)
 	return ch
